move evaluator tests to their own module
